This application has been cut down to just the core functionality that will be kept for a financial management hub.
Inventory has been removed. It may be added back
later in a much more feature rich environment but
for now the financial records the focus.

This works in both a CLI experience and in the
React GUI frontend.

Functionality is limited at this stage but will
be added as the project progresses.

Currently:
    Starting cash can be entered (cli only)
    Recurring expenses can be entered (cli only)
    Cash can be viewed
    Expenses can be viewed

More to come, soon.