The fundamental functions have been added to the engine to support most core financial transactions
The updates have been added to the command line
interface and are functional. These changes have
not been made to the React frontend yet but that
will be the next step.

The README has been updated but still has some of
the logic left from the previous version of the
program "business simulator". These will be removed
and updated for the next commit.

Still a rough draft at this point.